separation of concerns
separate items.js used in stocks check returns and dashboard from statuses.js which uses more advanced fetching
